# cleanup: speed up rm_DS_Store by multithread